backend/requirements.txt
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
mysqlclient==2.2.5
mysql-connector-python==9.1.0
python-dotenv==1.0.1
